Additional int to long changes in Comment service
Also fixes build problem by adding ArrayToCSV overload for a long[].

+++ Source/Podio .NET/Services/CommentService.cs
@@ -46,7 +46,7 @@ namespace PodioAPI.Services
         /// <param name="type"></param>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<List<Comment>> GetCommentsOnObject(string type, int id)
+        public async Task<List<Comment>> GetCommentsOnObject(string type, long id)
         {
             string url = string.Format("/comment/{0}/{1}/", type, id);
             return await _podio.Get<List<Comment>>(url);
@@ -108,7 +108,7 @@ namespace PodioAPI.Services
         /// </param>
         /// <param name="hook">todo: describe hook parameter on AddCommentToObject</param>
         /// <returns></returns>
-        public async Task<int> AddCommentToObject(string type, int id, CommentCreateUpdateRequest comment, bool alertInvite = false,
+        public async Task<int> AddCommentToObject(string type, long id, CommentCreateUpdateRequest comment, bool alertInvite = false,
             bool silent = false, bool hook = true)
         {
             string url = string.Format("/comment/{0}/{1}/", type, id);
